Cloud-Image-Galley/requirements.txt
boto3==1.24.91
botocore==1.27.91
click==8.1.3
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.9.0
APScheduler==3.9.1
mysql-connector-python==8.0.28
certifi==2021.10.8
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1